+++ api/instances/bundled-workbench-adapters.ts
@@ -1,0 +1,6 @@
+import { localWorkbenchAdapterRegistration } from './garbage/workbench.js'
+import type { LocalWorkbenchAdapterRegistration } from '../workbench/local-adapter-types'
+
+export const bundledLocalWorkbenchAdapterRegistrations: LocalWorkbenchAdapterRegistration[] = [
+  localWorkbenchAdapterRegistration,
+]