+++ api/instances/workbench-adapters.ts
@@ -1,6 +1,4 @@
-import { localWorkbenchAdapterRegistration } from './garbage/workbench.js'
+import { bundledLocalWorkbenchAdapterRegistrations } from './bundled-workbench-adapters.js'
 import type { LocalWorkbenchAdapterRegistration } from '../workbench/local-adapter-types'
 
-export const registeredLocalWorkbenchAdapters: LocalWorkbenchAdapterRegistration[] = [
-  localWorkbenchAdapterRegistration,
-]
+export const registeredLocalWorkbenchAdapters: LocalWorkbenchAdapterRegistration[] = bundledLocalWorkbenchAdapterRegistrations